Phase 1: real SFace recognizer, per-engine calibration, evidence-integrity fixes
Face identification core
- New `sface` engine (default via auto): OpenCV DNN YuNet detector with 5
  landmarks + landmark-aligned SFace 128-D embedding. No onnxruntime; SFace
  model auto-downloaded + SHA-256 pinned (`facechain fetch-models`).
  Fixtures: genuine pair ~0.98 cosine vs impostors <0.25 (LBPH was 0.90 / 0.6-0.8).
- `DetectedFace.landmarks`; factory preference insightface > sface > opencv > numpy.
- `face.calibration`: per-engine threshold + cosine->probability + match/likely/
  no-match band. `--threshold` now optional; resolved once the engine is known.

Notarised-evidence integrity (regressions from the merged live-search PR)
- `identity_guess/note/confidence` excluded from `record_hash` -> re-runs are
  byte-reproducible again (verified).
- `decided_by` gains `embedding_cosine+text_consensus` and is set truthfully when
  a scraped-name consensus overrides the top embedding score.
- Public probe hosting for SerpAPI Lens is now opt-in (`--allow-public-host` /
  `FACECHAIN_ALLOW_PUBLIC_PROBE_HOST`), off by default.

Housekeeping
- Remove stray nested `facechain-verify/requirements.txt`.
- Clear all pre-existing ruff + mypy(strict) errors from the merge; CI installs
  `[opencv]` and runs `fetch-models` so it exercises the real default engine.
- New tests: `test_sface_engine.py`, `test_trust_fixes.py`.

+++ src/facechain/enhance.py
@@ -32,7 +32,7 @@ def enhance_rgb_for_faces(rgb: np.ndarray) -> np.ndarray:
     if 0 < min_side < _MIN_SIDE_FOR_DETECT:
         scale = min(_MIN_SIDE_FOR_DETECT / float(min_side), _MAX_UPSCALE)
         img = img.resize(
-            (max(1, int(round(w * scale))), max(1, int(round(h * scale)))),
+            (max(1, round(w * scale)), max(1, round(h * scale))),
             resample=Image.Resampling.LANCZOS,
         )
